move spdlog_with_R.h into c++ tree
