+++ libtiledbsoma/include/tiledbsoma/spdlog_with_R.h
