+++ libtiledbsoma/src/logger.h
@@ -54,7 +54,7 @@
 #if !defined(R_BUILD)
 #include <spdlog/spdlog.h>
 #else
-#include <spdlog_with_R.h>
+#include "tiledbsoma/spdlog_with_R.h"
 #endif
 
 namespace tiledbsoma {
